requirements.txt
openai
python-dotenv
textual
rich
psutil
pydantic
requests
pyyaml
